+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Electrum plugin: Restore the previous `oneserver` setting when the plugin is disabled,
   to prevent users from inadvertently connecting to public Electrum servers with this setting still on.
 
+- Electrum plugin: allow specifying additional custom CLI arguments using the GUI
+
 ## 0.1.1 - 2020-05-27
 
 - Make bwt available as an Electrum plugin! 💥